+++ opencti-platform/opencti-front/src/private/components/riskAssessment/risks/RiskLine.js
@@ -16,12 +16,10 @@ import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
 import Checkbox from '@material-ui/core/Checkbox';
 import ListItemText from '@material-ui/core/ListItemText';
 import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
-import FiberManualRecordIcon from '@material-ui/icons/FiberManualRecord';
 import Skeleton from '@material-ui/lab/Skeleton';
 import inject18n from '../../../../components/i18n';
 import RiskAssessmentPopover from './RiskAssessmentPopover';
+import ItemIcon from '../../../../components/ItemIcon';
 
 const styles = (theme) => ({
   item: {
@@ -82,63 +80,7 @@ const styles = (theme) => ({
   chip: { borderRadius: '4px' },
 });
 
-const RiskTooltip = withStyles((theme) => ({
-  tooltip: {
-    backgroundColor: 'rgba(241, 241, 242, 0.25)',
-    color: '#FFF',
-    maxWidth: 220,
-    fontSize: theme.typography.pxToRem(12),
-    border: '1px solid rgba(241, 241, 242, 0.5)',
-    borderRadius: '4px',
-  },
-}))(Tooltip);
-
 class RiskLineComponent extends Component {
-  renderRiskLevels() {
-    const { node, t } = this.props;
-    const risk = node?.risk_level;
-    let level = 1;
-    let riskColor = '';
-
-    if (risk === 'very_high') {
-      level = 5;
-      riskColor = '#FC0D1B';
-      return Array.from({ length: level },
-        (item, index) => <RiskTooltip
-          title={node?.risk_level && t('Very High')}><IconButton style={{ padding: 0, minWidth: '1rem' }} key={index}><FiberManualRecordIcon style={{ fill: riskColor }}/></IconButton></RiskTooltip>);
-    }
-    if (risk === 'high') {
-      level = 4;
-      riskColor = '#F35426';
-      return Array.from({ length: level },
-        (item, index) => <RiskTooltip
-          title={node?.risk_level && t('High')}><IconButton style={{ padding: 0, minWidth: '1rem' }} key={index}><FiberManualRecordIcon style={{ fill: riskColor }}/></IconButton></RiskTooltip>);
-    }
-    if (risk === 'moderate') {
-      level = 3;
-      riskColor = '#E28120';
-      return Array.from({ length: level },
-        (item, index) => <RiskTooltip
-          title={node?.risk_level && t('Moderate')}><IconButton style={{ padding: 0, minWidth: '1rem' }} key={index}><FiberManualRecordIcon style={{ fill: riskColor }}/></IconButton></RiskTooltip>);
-    }
-    if (risk === 'low') {
-      level = 2;
-      riskColor = '#FFA800';
-      return Array.from({ length: level },
-        (item, index) => <RiskTooltip
-          title={node?.risk_level && t('Low')}><IconButton style={{ padding: 0, minWidth: '1rem' }} key={index}><FiberManualRecordIcon style={{ fill: riskColor }}/></IconButton></RiskTooltip>);
-    }
-    if (risk === 'very_low') {
-      level = 1;
-      riskColor = '#FCC434';
-      return Array.from({ length: level },
-        (item, index) => <RiskTooltip
-          title={node?.risk_level && t('Very Low')}><IconButton style={{ padding: 0, minWidth: '1rem' }} key={index}><FiberManualRecordIcon style={{ fill: riskColor }}/></IconButton></RiskTooltip>);
-    }
-
-    return <></>;
-  }
-
   render() {
     const {
       t,
@@ -220,7 +162,7 @@ class RiskLineComponent extends Component {
                   paddingLeft: '24px',
                 }}
               >
-                {this.renderRiskLevels()}
+                <ItemIcon variant='inline' type={node?.risk_level}/>
               </div>
               <div
                 className={classes.bodyItem}
